requirements.txt
langchain
google-generativeai
regex
pandas
